exec plan

+++ cmd/server/v1/handler/scenario-exec.go
@@ -17,7 +17,7 @@ type ScenarioExecCtrl struct {
 func (c *ScenarioExecCtrl) LoadExecData(ctx iris.Context) {
 	id, err := ctx.URLParamInt("id")
 
-	data, err := c.ScenarioExecService.LoadExecData(id)
+	data, err := c.ScenarioExecService.LoadExecData(uint(id))
 	if err != nil {
 		ctx.JSON(_domain.Response{Code: _domain.SystemErr.Code, Data: nil, Msg: err.Error()})
 		return
